+++ samples/request.php
@@ -1,6 +1,8 @@
 <?php
 
 use Piggly\ApiClient\Configuration;
+use Piggly\ApiClient\Exceptions\ApiRequestException;
+use Piggly\ApiClient\Exceptions\ApiResponseException;
 use Piggly\ApiClient\Request;
 
 require_once('../vendor/autoload.php');
@@ -9,5 +11,16 @@ $config = new Configuration();
 $config->host('https://jsonplaceholder.typicode.com/');
 
 $request = new Request($config);
-list($body, $code, $headers) = $request->get('/posts/{id}')->params(['id'=>1])->call();
+
+try {
+	$response = $request->get('/posts/{id}')->params(['id'=>1])->call();
+
+	$body = $response->getBody();
+	$status = $response->getStatus();
+	$headers = $response->getHeaders();
+} catch (ApiRequestException $e) {
+	$message = $e->getMessage();
+} catch (ApiResponseException $e) {
+	$response = $e->getResponse();
+}
 die();